refactor maps folder
