+++ lib/pages/home/home_page.dart
@@ -9,9 +9,9 @@ import 'package:trufi_core/blocs/home_page_cubit.dart';
 import 'package:trufi_core/blocs/preferences_cubit.dart';
 import 'package:trufi_core/widgets/fetch_error_handler.dart';
 
+import './plan_map/plan.dart';
+import './plan_map/plan_empty.dart';
 import '../../keys.dart' as keys;
-import '../../plan/plan.dart';
-import '../../plan/plan_empty.dart';
 import '../../trufi_app.dart';
 import '../../trufi_configuration.dart';
 import '../../trufi_models.dart';
