+++ lib/pages/home/plan_map/plan_map.dart
@@ -4,19 +4,17 @@ import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong/latlong.dart';
 import 'package:trufi_core/entities/plan_entity/plan_entity.dart';
+import 'package:trufi_core/widgets/map/buttons/crop_button.dart';
+import 'package:trufi_core/widgets/map/buttons/map_type_button.dart';
+import 'package:trufi_core/widgets/map/buttons/your_location_button.dart';
 import 'package:trufi_core/widgets/map/map_copyright.dart';
 import 'package:trufi_core/widgets/map/trufi_map_controller.dart';
+import 'package:trufi_core/widgets/map/trufi_online_map.dart';
 
-import '../composite_subscription.dart';
-import '../plan/plan.dart';
-import '../trufi_app.dart';
-import '../trufi_configuration.dart';
-import '../trufi_map_utils.dart';
-import '../widgets/map/buttons/crop_button.dart';
-import '../widgets/map/buttons/map_type_button.dart';
-import '../widgets/map/buttons/your_location_button.dart';
-import '../widgets/map/trufi_online_map.dart';
-
+import '../../../composite_subscription.dart';
+import '../../../trufi_app.dart';
+import '../../../trufi_map_utils.dart';
+import './plan.dart';
 const double customOverlayWidgetMargin = 80.0;
 
 typedef OnSelected = void Function(PlanItinerary itinerary);
@@ -81,7 +79,6 @@ class PlanMapPageState extends State<PlanMapPage>
 
   @override
   Widget build(BuildContext context) {
-    final cfg = TrufiConfiguration();
     final Locale locale = Localizations.localeOf(context);
     final theme = Theme.of(context);
     _data._selectedColor = theme.accentColor;
@@ -114,12 +111,11 @@ class PlanMapPageState extends State<PlanMapPage>
             ];
           },
         ),
-        if (cfg.map.satelliteMapTypeEnabled || cfg.map.terrainMapTypeEnabled)
-          Positioned(
-            top: 16.0,
-            right: 16.0,
-            child: _buildUpperActionButtons(context),
-          ),
+        const Positioned(
+          top: 16.0,
+          right: 16.0,
+          child: MapTypeButton(),
+        ),
         Positioned(
           bottom: 0,
           left: 0,
@@ -165,12 +161,6 @@ class PlanMapPageState extends State<PlanMapPage>
           ),
         )
       ],
-    );
-  }
-
-  Widget _buildUpperActionButtons(BuildContext context) {
-    return const SafeArea(
-      child: MapTypeButton(),
     );
   }
